Updated README.md
Added description and setup instructions.

I don't know if a separate repository for "new computer setup" should be created, that eventually installs this repository, or if the process can be fit into a single shell script in this repository.

With the variety of platforms, a separate repository will probably be the way to go. I already have experience with Ansible.

The software list in the README.md does not perfectly mirror the configuration files to be included in this repository, as it includes software I do use, but for which I use the default configuration.