Added prelimnary instructions for Windows install
Oh man, this was a doozy.

In this process, I learned enough to improve my already existing Windows
setup, and ironed out some its kinks.

Still not complete.

I'm not happy about wsl-ssh-pageant requiring a second window.
I would like it to work with the Start-Job comdlet in PowerShell
to make this happen, but couldn't nail it.

Still need to test this flow on a new computer to see if it actually
works as advertised.