Updated Windows documentation
We'll probably want to remerge the Windows documentation, because a lot
of the steps after getting gnupg installed are similar, except for
wsl-ssh-pageant.

Also, gpg --recv-key can take the fingerprint from gpg --card-status on
the signing key, and will grab the correct key, so no need to hard-code
the Fingerprint.

Oddly, though, gpg -K can be empty, and ssh-add -L will still grab
identities from the card just fine.

Start-Job does not appear to truly disconnect itself from a powershell
session (this may not be what it's meant to do), so the following does
work to create a process,

Start-Job -ScriptBlock {& wsl-ssh-pageant -systray -winssh "ssh-pageant"
-wsl (gpgconf --list-dirs agent-ssh-socket)}

But as soon as the console window is closed, the process exits.

All in all, a longterm goal would be to be able to git clone from the
https url, and then have the setup script adjust the reporitory remote
urlto the ssh url.