Started adding actions
It turns out dotdrop supports command execution, dynamic variables, and
loading (possibly dynamic) configuration files.

This is exactly what I need, and would eliminate the need for ansible or
something similar.

I just need to make sure that dotdrop calls the system shell, and so is
OS independent, and then which shell it calls on each device.

An alternative approach, which would be much simpler and easier to read,
would be to write a single-file app in python, assume python is
installed, and issue commands to that app that does everything in a
platform independent way, or tailored to each platform.