What should invoke be used for?
I think it'd be good to have invoke be used for only bootstrapping and
very complex tasks.

A lot of the tools I want to have installed, like plug.vim or oh-my-zsh,
already have curl/sh oneliners in their READMEs for easy installation.

These fit perfectly into dotdrop actions.

The invoke tool can be used to bootstrap the installation of those
tools, and to perform more complex tasks and installations.

Also, it is kind of cool that, since pipx uses pip, it can install from
anywhere pip can, like the git+https:// urls, or a local directory.

So I could put this tool in its own package in the repository, and have
it install itself with pipx, and make itself available everywhere.

Alternatively, if I can put it all in one file, I think that'd be much
better, as the tool would really only need to be used for this
repository, and hopefully won't be too complicated or long.

It seems to me like
setup.py --install=all --profile=""
would be an ideal thing for install.py to be able to run, to install all
the required tools used by dotdrop actions, and to start building a new
profile for the device we're running the command from.