Working on single-file invoke script
Should probably instead put .venv in a tempfolder with mktemp module,
because git clone won't work if anything is in the folder already

Iv'e heard programs may be able to run from unlinked folders, so that
cleanup is easier: create dir, install files, run and unlink them.

It will be tough to figure out how to reliably uninstall unwanted
versions of Python on Windows, but it may be enough to change the PATH
ordering.

I think it'd be good to also create a gist for a
singlefile_invoke_script.py template

Since this appears to be kind of in a working state, moving forward,
I want to develop the required functionality, while also adding logging
and using Python's auditing framework to track file access.
One of the things that's burned me a lot has been accidentally
overwriting files. Now that I can have an interactive script, I don't
want to unwittingly overwrite something.

When I do implement logging, I want to model it how invoke does
logging, utilizing the logging module, so that logging calls are
debug(), error(), etc.

The script also needs to be reorganized.