fix(chezmoi/bootstrap): remove curl since nu has http get
I would have left it, but it seems like it has to be explicitly run
through /bin/sh on linux, and while it can be assimilated and the runs
fine, it doesn't seem to assimilate correctly when done with `/bin/sh
curl --assimilate`, only `/bin/sh assimilate curl`, which requires the
second executable, `assimilate`, and is quite a few extra steps. Those
steps can be done in a nu script if they're needed.