fix(nvim/telescope): merge in telescope-undo
Also copy some keybindings from:

https://github.com/Olical/dotfiles/blob/0fb7300cac6bdc2a2d894dbda448edeb6a10fc97/stowed/.config/nvim/fnl/plugins/telescope.fnl#L20-L41

Also remove redundant keybindings: anything defined in the lazy.nvim
config is also picked up by which-key, so no need to redefine it in a
custom config function.

Also, switch to using the suggested way of organizing telescope plugins:
have Telescope load them in a custom config function on that plugin, put
the configuration into the main Telescope config, and add plugin
specific keybindings into that plugin's lazy.nvim keybindings, so that
that plugin is only loaded when the keybindings that require its
functionality are activated.